Refaktorerer ut ArbedisgiverDetaljer med tilhørende klasser og validering til egen fil.
