+++ src/main/kotlin/no/nav/helse/soknad/Søknad.kt
@@ -63,10 +63,6 @@ enum class BarnRelasjon(val utskriftsvennlig: String) {
     ANNET("Annet")
 }
 
-data class ArbeidsgiverDetaljer(
-    val organisasjoner: List<OrganisasjonDetaljer>
-)
-
 data class Medlemskap(
     val harBoddIUtlandetSiste12Mnd: Boolean? = null,
     val utenlandsoppholdSiste12Mnd: List<Bosted> = listOf(),
@@ -78,22 +74,6 @@ data class UtenlandsoppholdIPerioden(
     val skalOppholdeSegIUtlandetIPerioden: Boolean? = null,
     val opphold: List<Utenlandsopphold> = listOf()
 )
-
-data class OrganisasjonDetaljer(
-    val navn: String? = null,
-    val skalJobbe: String,
-    val organisasjonsnummer: String,
-    val jobberNormaltTimer: Double,
-    val skalJobbeProsent: Double,
-    val vetIkkeEkstrainfo: String? = null,
-    val arbeidsform: Arbeidsform? = null //TODO 09.02.2021 - Fjerne nullable når prodsatt. Påbudt felt
-)
-
-enum class Arbeidsform{
-    FAST,
-    TURNUS,
-    VARIERENDE
-}
 
 enum class TilsynsordningSvar {
     ja,
